siguldry-client: Split up the mega-module a bit
Adjust the client module to match the server and bridge layouts. There
are no functional changes here.
